requirements.txt
streamlit
pandas
plotly
wordcloud
pillow 